+++ src/shared/hooks/index.ts
@@ -1,2 +1,3 @@
+export { useQueryString } from './useQueryString';
 export { useShowErrorToast } from './useShowErrorToast';
 export { useTimer } from './useTimer';
